[IMP] web: convert core/file_upload tests to hoot
task-id: 3705027
